+++ app/sprinkles/admin/tests/Integration/SprunjeTests.php
@@ -4,17 +4,17 @@ namespace UserFrosting\Tests\Integration;
 
 use UserFrosting\Sprinkle\Admin\Sprunje\UserPermissionSprunje;
 use UserFrosting\Sprinkle\Core\Util\ClassMapper;
-use UserFrosting\Tests\DatabaseTransactions;
 use UserFrosting\Tests\TestCase;
 use UserFrosting\Tests\WithTestDatabase;
+use UserFrosting\Tests\RefreshDatabase;
 
 /**
  * Integration tests for the built-in Sprunje classes.
  */
 class SprunjeTests extends TestCase
 {
-    use DatabaseTransactions;
     use WithTestDatabase;
+    use RefreshDatabase;
 
     /**
      *    @var ClassMapper
@@ -31,9 +31,6 @@ class SprunjeTests extends TestCase
         parent::setUp();
 
         $this->classMapper = new ClassMapper();
-
-        // Setup the db
-        $this->ci->migrator->run();
     }
 
     /**